+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/converter/YamlDockerStepConverter.java
@@ -15,7 +15,15 @@ public class YamlDockerStepConverter implements StepConverter<YamlDockerStep> {
         Chunk c = new Chunk();
 
         String id = ctx.nextId();
-        ELCall call = createELCall("docker", Arrays.asList(s.getImage(), s.isForcePull(), s.getCmd(), s.getEnv(), "${" + InternalConstants.Context.LOCAL_PATH_KEY + "}"));
+
+        Object args = Arrays.asList(s.getImage(),
+                s.isForcePull(),
+                s.isDebug(),
+                s.getCmd(),
+                s.getEnv(),
+                "${" + InternalConstants.Context.LOCAL_PATH_KEY + "}");
+        ELCall call = createELCall("docker", args);
+
         c.addElement(new ServiceTask(id, ExpressionType.SIMPLE, call.getExpression(), call.getArgs(), null, true));
         c.addOutput(id);
         c.addSourceMap(id, toSourceMap(s, "Docker: " + s.getImage()));